Make overflow test more robust
Test with large multiplier

Change multiplier type from int to long long

+++ mixed_congruential.cpp
@@ -1,7 +1,7 @@
 #include "linear_congruential.h"
 #include "mixed_congruential.h"
 
-MixedCongruentialGenerator::MixedCongruentialGenerator(long long modulus, int multiplier, int increment) :
+MixedCongruentialGenerator::MixedCongruentialGenerator(long long modulus, long long multiplier, long long increment) :
 	LinearCongruentialGenerator(modulus, 0), multiplier(multiplier), increment(increment) {};
 
 void MixedCongruentialGenerator::updateState() {
